# [master < T591-FL] Enhance performance of Property Store on node creation