--- a/src/query/common.hpp
+++ b/src/query/common.hpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Memgraph Ltd.
+// Copyright 2023 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -104,6 +104,39 @@ storage::PropertyValue PropsSetChecked(T *record, const storage::PropertyId &key
     return std::move(*maybe_old_value);
   } catch (const TypedValueException &) {
     throw QueryRuntimeException("'{}' cannot be used as a property value.", value.type());
+  }
+}
+
+template <typename T>
+concept AccessorWithSetProperties = requires(T accessor,
+                                             const std::map<storage::PropertyId, storage::PropertyValue> &properties) {
+  { accessor.SetProperties(properties) } -> std::same_as<storage::Result<std::vector<storage::PropertyValue>>>;
+};
+
+/// Set a property `value` mapped with given `key` on a `record`.
+///
+/// @throw QueryRuntimeException if value cannot be set as a property value
+template <AccessorWithSetProperty T>
+storage::PropertyValue MultiPropsSetChecked(T *record,
+                                            std::map<storage::PropertyId, storage::PropertyValue> &properties) {
+  try {
+    auto maybe_old_value = record->SetProperties(properties);
+    if (maybe_old_value.HasError()) {
+      switch (maybe_old_value.GetError()) {
+        case storage::Error::SERIALIZATION_ERROR:
+          throw TransactionSerializationException();
+        case storage::Error::DELETED_OBJECT:
+          throw QueryRuntimeException("Trying to set properties on a deleted object.");
+        case storage::Error::PROPERTIES_DISABLED:
+          throw QueryRuntimeException("Can't set property because properties on edges are disabled.");
+        case storage::Error::VERTEX_HAS_EDGES:
+        case storage::Error::NONEXISTENT_OBJECT:
+          throw QueryRuntimeException("Unexpected error when setting a property.");
+      }
+    }
+    return std::move(*maybe_old_value);
+  } catch (const TypedValueException &) {
+    throw QueryRuntimeException("Cannot set properties.");
   }
 }