--- a/src/storage/v2/edge_accessor.cpp
+++ b/src/storage/v2/edge_accessor.cpp
@@ -123,7 +123,7 @@ Result<storage::PropertyValue> EdgeAccessor::SetProperty(PropertyId property, co
   return std::move(current_value);
 }
 
-Result<bool> EdgeAccessor::SetProperties(const std::map<storage::PropertyId, storage::PropertyValue> &properties) {
+Result<bool> EdgeAccessor::InitProperties(const std::map<storage::PropertyId, storage::PropertyValue> &properties) {
   utils::MemoryTracker::OutOfMemoryExceptionEnabler oom_exception;
   if (!config_.properties_on_edges) return Error::PROPERTIES_DISABLED;
 
@@ -133,7 +133,7 @@ Result<bool> EdgeAccessor::SetProperties(const std::map<storage::PropertyId, sto
 
   if (edge_.ptr->deleted) return Error::DELETED_OBJECT;
 
-  if (!edge_.ptr->properties.SetProperties(properties)) return false;
+  if (!edge_.ptr->properties.InitProperties(properties)) return false;
   for (const auto &[property, _] : properties) {
     CreateAndLinkDelta(transaction_, edge_.ptr, Delta::SetPropertyTag(), property, PropertyValue());
   }