--- a/src/storage/v2/edge_accessor.hpp
+++ b/src/storage/v2/edge_accessor.hpp
@@ -61,7 +61,7 @@ class EdgeAccessor final {
   /// Set property values only if property store is empty. Returns `true` if successully set all values,
   /// `false` otherwise.
   /// @throw std::bad_alloc
-  Result<bool> SetProperties(std::map<storage::PropertyId, storage::PropertyValue> &properties);
+  Result<bool> SetProperties(const std::map<storage::PropertyId, storage::PropertyValue> &properties);
 
   /// Remove all properties and return old values for each removed property.
   /// @throw std::bad_alloc