--- a/src/storage/v2/property_store.hpp
+++ b/src/storage/v2/property_store.hpp
@@ -61,6 +61,10 @@ class PropertyStore {
   /// @throw std::bad_alloc
   bool SetProperty(PropertyId property, const PropertyValue &value);
 
+  /// Set property values and return `true` if insertion took place. `false` is
+  /// returned if there exists property in property store and insertion couldn't take place. The time complexity of this
+  /// function is O(n).
+  /// @throw std::bad_alloc
   bool SetProperties(std::map<storage::PropertyId, storage::PropertyValue> &properties);
 
   /// Remove all properties and return `true` if any removal took place.