… src/storage/v2/vertex_accessor.cpp
… src/storage/v2/vertex_accessor.hpp
… tests/gql_behave/tests/memgraph_V1/features/functions.feature
… tests/unit/storage_v2_property_store.cpp