[master < Task] PR
- [x] Check, and update documentation if necessary

**Release note**: Improve performance by storing node or edge properties in a property store all at once on node or edge creation, instead of individually,  when dealing with large numbers of properties.

**Benchmarking**
- [x] Import 1 000 000 lines, from each line create 2 nodes and 1 new edge. Each line will have X properties, of 10 string length. Each node has X/2 features, where X is even.

| Number of properties per line | Import time (OLD) [seconds] | Import time (NEW) [seconds] | Throughput (OLD) [lines/s] | Throughput (NEW) [lines/s] |
| -- | -- | -- |  -- | -- |
|2 | 3.6s | 3.6 s|  |  |
|10 | 9.6s | 8.7 s|   |  |
|20 | 22s | 17.5 s|  |   |